fix: srcObject issue
