+++ client/src/hooks/useVideoCall.jsx
@@ -1,33 +1,40 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState, useContext } from "react";
 import Peer from "simple-peer";
 import io from "socket.io-client";
+import { MyDataContext } from "../pages/Home";
+import { MyGlobalContext } from "../context/MyGlobalContext";
 
 const socket = io.connect("http://localhost:3001");
 
-const useVideoChat = () => {
-  const [me, setMe] = useState("");
-  const [stream, setStream] = useState();
+export function useVideoCall() {
+  const [stream, setStream] = useState("");
   const [receivingCall, setReceivingCall] = useState(false);
   const [caller, setCaller] = useState("");
   const [callerSignal, setCallerSignal] = useState();
   const [callAccepted, setCallAccepted] = useState(false);
-  const [idToCall, setIdToCall] = useState("");
   const [callEnded, setCallEnded] = useState(false);
-  const [name, setName] = useState("");
+  const { roomId, _id, name, setName, video, setVideo, audio, setAudio } =
+    useContext(MyGlobalContext);
+  const { contactData } = useContext(MyDataContext);
   const myVideo = useRef();
-  const userVideo = useRef();   
+  const userVideo = useRef();
   const connectionRef = useRef();
 
   useEffect(() => {
     navigator.mediaDevices
-      .getUserMedia({ video: true, audio: true })
+      .getUserMedia({ video: video, audio: audio })
       .then((stream) => {
         setStream(stream);
-        myVideo.current.srcObject = stream;
+        if (myVideo.current) {
+          myVideo.current.srcObject = stream;
+        }
       });
 
-    socket.on("me", (id) => {
-      setMe(id);
+    socket.emit("join_room", {
+      roomId: roomId,
+      userId: _id,
+      name: name,
+      message: "Video Call",
     });
 
     socket.on("callUser", (data) => {
@@ -36,26 +43,33 @@ const useVideoChat = () => {
       setName(data.name);
       setCallerSignal(data.signal);
     });
+
+    return () => {
+      socket.off("callUser");
+    };
   }, []);
 
-  const callUser = (id) => {
-    console.log("call user");
+  const callUser = () => {
     const peer = new Peer({
       initiator: true,
       trickle: false,
       stream: stream,
     });
+
     peer.on("signal", (data) => {
       socket.emit("callUser", {
-        userToCall: id,
+        userToCall: contactData._id,
         signalData: data,
-        from: me,
+        from: _id,
         name: name,
+        roomId: roomId,
       });
     });
+
     peer.on("stream", (stream) => {
       userVideo.current.srcObject = stream;
     });
+
     socket.on("callAccepted", (signal) => {
       setCallAccepted(true);
       peer.signal(signal);
@@ -65,16 +79,21 @@ const useVideoChat = () => {
   };
 
   const answerCall = () => {
-    console.log("answer call")
     setCallAccepted(true);
     const peer = new Peer({
       initiator: false,
       trickle: false,
       stream: stream,
     });
+
     peer.on("signal", (data) => {
-      socket.emit("answerCall", { signal: data, to: caller });
+      socket.emit("answerCall", {
+        signal: data,
+        to: caller,
+        roomId: roomId,
+      });
     });
+
     peer.on("stream", (stream) => {
       userVideo.current.srcObject = stream;
     });
@@ -84,36 +103,24 @@ const useVideoChat = () => {
   };
 
   const leaveCall = () => {
-    console.log("leave call");
     setCallEnded(true);
-    connectionRef.current.destroy();
-  };
-
-  const copyIdToClipboard = () => {
-    navigator.clipboard.writeText(me);
+    if (callEnded) {
+      connectionRef.current.destroy();
+    }
   };
 
   return {
-    me,
-    stream,
-    receivingCall,
-    caller,
-    callerSignal,
-    callAccepted,
-    idToCall,
-    callEnded,
-    name,
     myVideo,
     userVideo,
-    connectionRef,
-    setIdToCall,
-    setName,
+    stream,
+    receivingCall,
+    callAccepted,
+    callEnded,
     callUser,
     answerCall,
-    setCallAccepted,
     leaveCall,
-    copyIdToClipboard,
+    name,
+    video,
+    setReceivingCall,
   };
-};
-
-export default useVideoChat;
+}
