--- server/src/controllers/socketController copy.js
@@ -1,102 +1,0 @@
-// let io;
-
-// function initialize(server) {
-//   io = require("socket.io")(server, {
-//     cors: {
-//       origin: process.env.CORS_ORIGIN || `http://localhost:3000`,
-//       // methods: ['GET', 'POST']
-//     },
-//   });
-
-//   io.on("connection", (socket) => {
-//     console.log(`User Connected: ${socket.id}`);
-
-//     socket.on("join_room", (data) => {
-//       socket.join(data);
-//       console.log(`User with ID: ${socket.id} joined room: ${data}`);
-//     });
-
-//     socket.on("send_message", (data) => {
-//       socket.to(data.room).emit("receive_message", data);
-//     });
-
-//     socket.on("start_call", (data) => {
-//       socket.to(data.room).emit("call_started", data);
-//     });
-
-//     socket.on("answer_call", (data) => {
-//       socket.to(data.room).emit("call_answered", data);
-//     });
-
-//     socket.on("reject_call", (data) => {
-//       socket.to(data.room).emit("call_rejected", data);
-//     });
-
-//     socket.on("end_call", (data) => {
-//       socket.to(data.room).emit("call_ended", data);
-//     });
-
-//     socket.on("disconnect", () => {
-//       socket.broadcast.emit(socket.id +"left the room")
-//       console.log("User Disconnected", socket.id);
-//     });
-//   });
-// }
-
-// module.exports = {
-//   initialize,
-// };
-
-
-let io;
-
-function initialize(server) {
-  io = require("socket.io")(server, {
-    cors: {
-      origin: process.env.CORS_ORIGIN || `http://localhost:3000`,
-      methods: ["GET", "POST"],
-    },
-  });
-
-  io.on("connection", (socket) => {
-    socket.on("join_room", (data) => {
-      socket.join(data.roomId);
-      console.log(`User: "${data.name}" | Joined the room: ${data.roomId} | ${data.message}`);
-    });
-
-    socket.on("send_message", (data) => {
-      console.log(data);
-      socket.to(data.room).emit("receive_message", data);
-    });
-
-    socket.emit("me", socket.id);
-
-    socket.on("callUser", (data) => {
-      io.to(data.roomId).emit("callUser", {
-        signal: data.signalData,
-        from: data.from,
-        name: data.name,
-        userToCall: data.userToCall,
-      });
-      console.log(data);
-    });
-    
-
-    socket.on("answerCall", (data) => {
-      io.to(data.to).emit("callAccepted", data.signal);
-    });
-   
-
-    socket.on("disconnect", () => {
-      socket.rooms.forEach((room) => {
-        socket.to(room).emit("callEnded");
-        socket.leave(room);
-      });
-    });
-    
-  });
-}
-
-module.exports = {
-  initialize,
-};
